pyproject.toml
[project]
name = "sre-bench"
version = "1.0.0"
description = "Autonomous incident response benchmark for SRE agents."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn>=0.29.0",
    "pydantic>=2.7.0",
    "numpy>=1.26.0",
    "openai>=1.0.0",
    "httpx>=0.27.0",
    "openenv-core>=0.2.0",
    "python-dotenv>=1.0.0"
]

[project.scripts]
server = "server.app:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
